+++ src/models/constants.ts
@@ -10,4 +10,10 @@ export const defaultAmountToDarkenLighten = 10;
 
 export const defaultSaturation = 0.5;
 
+export const azureBlue = '#007fff';
 export const peacockGreen = '#42b883';
+
+export const peacockMementos = {
+  favoritesVersion: `${extensionShortName}.favoritesVersion`,
+  peacockColor: `${extensionShortName}.peacockColor`
+};