+++ src/remote/enums.ts
@@ -3,3 +3,9 @@ export enum RemoteCommands {
   changeColorOfRemoteSsh = 'peacock.changeColorOfRemoteSsh',
   changeColorOfRemoteContainers = 'peacock.changeColorOfRemoteContainers'
 }
+
+export enum RemoteNames {
+  wsl = 'wsl',
+  sshRemote = 'ssh-remote',
+  devContainer = 'dev-container'
+}